Fix #1210 - Improve interceptors to get path parameter related details
With this commit request.getPathParams() method is intorduced to java interceptors to get path parameter details.
And for ballerina interceptors users can get it as invocationContext.attributes["pathParams"] and cast it to a map.
With this commit users can details api_name, api_version, matching_resource , using this properties from the invocation context as well.