This commit brings back the blocked APIs lifecycle state managment to the choreo branch. The /apis call will be only made during the startup. Then onwards the JMS/ASB events are used. APIMetadataMap internally stores the api metadata information received from the /apis REST api response and via events.
The API MetadataMap only stores the metadata either it is a blockedAPI or else if it is default versioned API. (in choreo branch, default versioned API is still not supported). If the apimetadata object is updated such that it does not satisfy either one these conditions, those would be removed.

From now on, the APIListDiscoverty Client is no longer in use. And SubscriptionDataStore's API object is also remain unused. The lifecycle state is stored within the APIFactory's APIs which represent the runtime-artifacts. Hence the requirement of keeping two proto structures to maintain, is reduced to single structure.

PrototypedEndpoints Testcase is removed as the feature is no longer available in APIM. Hence it is not required to support.